donation_project/requirements.txt
APScheduler==3.10.4
asgiref==3.9.1
bcrypt==4.1.2
blinker==1.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.7
colorama==0.4.6
Django==5.2.4
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.5
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.5.3
stripe==12.5.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.5.0
Werkzeug==3.0.1
wheel==0.45.1
WTForms==3.1.2
